feat: Reorganize layouts view
