+++ lib/models/device.dart
@@ -25,7 +25,7 @@ import 'package:bluecherry_client/providers/server_provider.dart';
 import 'package:bluecherry_client/providers/settings_provider.dart';
 import 'package:bluecherry_client/utils/config.dart';
 import 'package:bluecherry_client/utils/extensions.dart';
-import 'package:bluecherry_client/widgets/device_grid/desktop/external_stream.dart';
+import 'package:bluecherry_client/screens/layouts/desktop/external_stream.dart';
 import 'package:flutter/foundation.dart';
 
 class ExternalDeviceData {
