+++ lib/screens/layouts/desktop/device_info_dialog.dart
