+++ lib/screens/layouts/desktop/stream_data.dart
@@ -21,9 +21,9 @@ import 'dart:async';
 
 import 'package:bluecherry_client/models/device.dart';
 import 'package:bluecherry_client/providers/settings_provider.dart';
+import 'package:bluecherry_client/screens/layouts/desktop/external_stream.dart';
 import 'package:bluecherry_client/utils/config.dart';
 import 'package:bluecherry_client/utils/extensions.dart';
-import 'package:bluecherry_client/widgets/device_grid/desktop/external_stream.dart';
 import 'package:bluecherry_client/widgets/ptz.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
