+++ lib/screens/layouts/device_grid.dart
@@ -29,6 +29,13 @@ import 'package:bluecherry_client/providers/home_provider.dart';
 import 'package:bluecherry_client/providers/mobile_view_provider.dart';
 import 'package:bluecherry_client/providers/server_provider.dart';
 import 'package:bluecherry_client/providers/settings_provider.dart';
+import 'package:bluecherry_client/screens/layouts/desktop/device_info_dialog.dart';
+import 'package:bluecherry_client/screens/layouts/desktop/external_stream.dart';
+import 'package:bluecherry_client/screens/layouts/desktop/layout_manager.dart';
+import 'package:bluecherry_client/screens/layouts/desktop/multicast_view.dart';
+import 'package:bluecherry_client/screens/layouts/desktop/stream_data.dart';
+import 'package:bluecherry_client/screens/layouts/mobile/device_view.dart';
+import 'package:bluecherry_client/screens/layouts/video_status_label.dart';
 import 'package:bluecherry_client/utils/app_links/app_links.dart' as app_links;
 import 'package:bluecherry_client/utils/constants.dart';
 import 'package:bluecherry_client/utils/extensions.dart';
@@ -38,13 +45,6 @@ import 'package:bluecherry_client/utils/video_player.dart';
 import 'package:bluecherry_client/utils/widgets/squared_icon_button.dart';
 import 'package:bluecherry_client/utils/window.dart';
 import 'package:bluecherry_client/widgets/collapsable_sidebar.dart';
-import 'package:bluecherry_client/widgets/device_grid/desktop/device_info_dialog.dart';
-import 'package:bluecherry_client/widgets/device_grid/desktop/external_stream.dart';
-import 'package:bluecherry_client/widgets/device_grid/desktop/layout_manager.dart';
-import 'package:bluecherry_client/widgets/device_grid/desktop/multicast_view.dart';
-import 'package:bluecherry_client/widgets/device_grid/desktop/stream_data.dart';
-import 'package:bluecherry_client/widgets/device_grid/mobile/device_view.dart';
-import 'package:bluecherry_client/widgets/device_grid/video_status_label.dart';
 import 'package:bluecherry_client/widgets/error_warning.dart';
 import 'package:bluecherry_client/widgets/hover_button.dart';
 import 'package:bluecherry_client/widgets/misc.dart';
