+++ lib/screens/servers/additional_server_settings.dart
@@ -19,9 +19,9 @@
 
 import 'package:bluecherry_client/models/server.dart';
 import 'package:bluecherry_client/providers/settings_provider.dart';
+import 'package:bluecherry_client/screens/layouts/desktop/stream_data.dart';
 import 'package:bluecherry_client/screens/servers/wizard.dart';
 import 'package:bluecherry_client/utils/methods.dart';
-import 'package:bluecherry_client/widgets/device_grid/desktop/stream_data.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:provider/provider.dart';
