+++ lib/widgets/events_timeline/desktop/timeline.dart
@@ -29,7 +29,7 @@ import 'package:bluecherry_client/utils/extensions.dart';
 import 'package:bluecherry_client/utils/methods.dart';
 import 'package:bluecherry_client/utils/widgets/squared_icon_button.dart';
 import 'package:bluecherry_client/widgets/collapsable_sidebar.dart';
-import 'package:bluecherry_client/widgets/device_grid/device_grid.dart'
+import 'package:bluecherry_client/screens/layouts/device_grid.dart'
     show calculateCrossAxisCount;
 import 'package:bluecherry_client/widgets/events/events_screen.dart';
 import 'package:bluecherry_client/widgets/events_timeline/desktop/timeline_card.dart';
