+++ lib/widgets/multi_window/single_layout_window.dart
@@ -19,7 +19,7 @@
 
 import 'package:bluecherry_client/models/layout.dart';
 import 'package:bluecherry_client/widgets/desktop_buttons.dart';
-import 'package:bluecherry_client/widgets/device_grid/device_grid.dart';
+import 'package:bluecherry_client/screens/layouts/device_grid.dart';
 import 'package:flutter/material.dart';
 
 class AlternativeLayoutView extends StatelessWidget {
