+++ lib/widgets/settings/desktop/server.dart
@@ -19,7 +19,7 @@
 
 import 'package:bluecherry_client/providers/settings_provider.dart';
 import 'package:bluecherry_client/utils/extensions.dart';
-import 'package:bluecherry_client/widgets/device_grid/video_status_label.dart';
+import 'package:bluecherry_client/screens/layouts/video_status_label.dart';
 import 'package:bluecherry_client/widgets/settings/desktop/settings.dart';
 import 'package:bluecherry_client/widgets/settings/mobile/settings.dart';
 import 'package:flutter/foundation.dart';
